handle java rename issue individually
